refactor: extensive code refactoring to address multiple tracked issues
