+++ src/modules/assessment/domain/question-repository.interface.ts
@@ -1,4 +1,5 @@
-import { Question } from './form.entity';
+import { Question } from "./question.entity";
+
 
 export interface QuestionRepository {
   findByFormId(formId: number): Promise<Question[]>;
